BigInteger -> long

+++ src/main/Main.java
@@ -5,14 +5,12 @@ import encryption.EncryptionService;
 import input.InputHandler;
 import keygen.KeyGenService;
 
-import java.math.BigInteger;
-
 public class Main {
 
     private static Tasks chooseTask() {
         System.out.println("Milyen műveletet szeretne végezni? (titkosítás: 1, visszafejtés: 2, kulcsgenerálás: 3)");
-        BigInteger taskNum = InputHandler.getTaskInput("Művelet");
-        return Tasks.values()[taskNum.intValue() - 1];
+        long taskNum = InputHandler.getTaskInput("Művelet");
+        return Tasks.values()[(int) (taskNum - 1)];
     }
 
     public static void main(String[] args) {
